Introduce transfer modal component
